Add files via upload
fixed the subfunction theta to account for the varying angle of incidence between camera and FOV. also made it based on relative FOV location and not the global coordinate system. 
Added mutliple search boxes with random sizes to find the best line of fit and get rid of dependencies on location of the search box.
Added mean subtraction, edge function works much better now.